70/100%

+++ src/main/java/com/demo_banking/models/User.java
@@ -9,18 +9,18 @@ import java.util.Date;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long user_id;
+    private int user_id;
     private String first_name;
     private String last_name;
     private String email;
     private String password;
     private Date data_created;
 
-    public Long getUser_id() {
+    public int getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(Long user_id) {
+    public void setUser_id(int user_id) {
         this.user_id = user_id;
     }
